pyproject.toml
[tool.poetry]
name = "twikit-crawler"
version = "0.1.0"
description = ""
authors = ["ducnguyen-theinfitech <duc_nguyen@theinfitech.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
twikit = "^2.1.3"
playwright = "^1.49.0"
tqdm = "^4.67.1"
boto3 = "^1.35.68"
psycopg2-binary = "^2.9.10"
xvfbwrapper = "^0.2.9"
sqlalchemy = "^2.0.36"
pydantic = "^2.10.1"
python-dotenv = "^1.0.1"
pandas = "^2.2.3"
numpy = "^1.26"
pyspark = "^3.5.3"
delta-spark = "^3.2.1"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
